+++ component/admin/controllers/icalrepeat.php
@@ -12,6 +12,8 @@ defined('_JEXEC') or die('Restricted access');
 
 jimport('joomla.application.component.controller');
 
+use Joomla\Utilities\ArrayHelper;
+
 class AdminIcalrepeatController extends JControllerLegacy
 {
 
@@ -50,7 +52,7 @@ class AdminIcalrepeatController extends JControllerLegacy
 		$db = JFactory::getDBO();
 		$publishedOnly = false;
 		$cid = $jinput->get('cid', array(0),"array");
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 
 		if (is_array($cid) && count($cid) > 0)
 			$id = $cid[0];
@@ -146,7 +148,7 @@ class AdminIcalrepeatController extends JControllerLegacy
 
 		$db = JFactory::getDBO();
 		$cid = JRequest::getVar('cid', array(0));
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		if (is_array($cid) && count($cid) > 0)
 			$id = $cid[0];
 		else
@@ -540,7 +542,7 @@ class AdminIcalrepeatController extends JControllerLegacy
 			$detail->evdet_id = $rpt->eventdetail_id;
 		}
 
-		$detail->priority = intval(JArrayHelper::getValue($array, "priority", 0));
+		$detail->priority = intval(ArrayHelper::getValue($array, "priority", 0));
 
 		$detail->store();
 
@@ -727,7 +729,7 @@ class AdminIcalrepeatController extends JControllerLegacy
 		$cid = JRequest::getVar('cid', array(0));
 		if (!is_array($cid))
 			$cid = array(intval($cid));
-		JArrayHelper::toInteger($cid);
+			ArrayHelper::toInteger($cid);
 
 		$db = JFactory::getDBO();
 		foreach ($cid as $id)
@@ -842,7 +844,7 @@ class AdminIcalrepeatController extends JControllerLegacy
 		$cid = JRequest::getVar('cid', array(0));
 		if (!is_array($cid))
 			$cid = array(intval($cid));
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 
 		$db = JFactory::getDBO();
 		foreach ($cid as $id)