+++ component/admin/controllers/icals.php
@@ -13,6 +13,7 @@ defined( '_JEXEC' ) or die( 'Restricted access' );
 
 jimport('joomla.application.component.controllerform');
 
+use Joomla\Utilities\ArrayHelper;
 
 class AdminIcalsController extends JControllerForm {
 
@@ -156,7 +157,7 @@ class AdminIcalsController extends JControllerForm {
 		$this->view = $this->getView("icals","html");
 
 		$cid	= JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		if (is_array($cid) && count($cid)>0) $editItem=$cid[0];
 		else $editItem=0;
 
@@ -247,7 +248,7 @@ class AdminIcalsController extends JControllerForm {
 			return;
 		}
 		$cid	= JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		if (is_array($cid) && count($cid)>0) {
 			$cid=$cid[0];
 		} else {
@@ -394,7 +395,7 @@ class AdminIcalsController extends JControllerForm {
 
 		$icsid = intval(JRequest::getVar('icsid',0));
 		$cid	= JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		if (is_array($cid) && count($cid)>0) {
 			$cid=$cid[0];
 		} else {
@@ -465,13 +466,13 @@ class AdminIcalsController extends JControllerForm {
 
 	function publish(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleICalPublish($cid,1);
 	}
 
 	function unpublish(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleICalPublish($cid,0);
 	}
 
@@ -495,13 +496,13 @@ class AdminIcalsController extends JControllerForm {
 
 	function autorefresh(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleAutorefresh($cid,1);
 	}
 
 	function noautorefresh(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleAutorefresh($cid,0);
 	}
 
@@ -525,13 +526,13 @@ class AdminIcalsController extends JControllerForm {
 
 	function isdefault(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleDefault($cid,1);
 	}
 
 	function notdefault(){
 		$cid = JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 		$this->toggleDefault($cid,0);
 	}
 
@@ -597,7 +598,7 @@ class AdminIcalsController extends JControllerForm {
 		JRequest::checkToken() or jexit( 'Invalid Token' );
 
 		$cid	= JRequest::getVar(	'cid',	array(0) );
-		JArrayHelper::toInteger($cid);
+		ArrayHelper::toInteger($cid);
 
 		$db	= JFactory::getDBO();