+++ component/admin/tables/jevuser.php
@@ -12,6 +12,8 @@
 // no direct access
 defined('_JEXEC') or die('Restricted access');
 
+use Joomla\Utilities\ArrayHelper;
+
 /**
 * User Table class
 *
@@ -89,13 +91,13 @@ class TableUser extends JTable
 		$join = array();
 		if (is_array($ids)){
 			if (count($ids)>0){
-				JArrayHelper::toInteger($ids);
+				ArrayHelper::toInteger($ids);
 				$idstring = implode(",",$ids);
 				$where[] = " tl.id in ($idstring)";
 			}
 		}
 		else {
-			$idsstring = intval($ids);
+			$idstring = intval($ids);
 			$where[] = "tl.id in ($idstring)";
 		}
 
@@ -182,7 +184,7 @@ class TableUser extends JTable
 
 	public static function getUsersByUserid($userid,$index="id"){
 		if (is_array($userid)){
-			JArrayHelper::toInteger($userid);
+			ArrayHelper::toInteger($userid);
 			$userids = implode(",",$userid);
 		}
 		else {
@@ -286,14 +288,14 @@ class TableUser extends JTable
 		if (key_exists('categories', $array)) {
 			if($array['categories']=='all' || $array['categories']=='none') $this->categories = $array['categories'];
 			else if (is_array($array['categories'])){
-				JArrayHelper::toInteger($array['categories']);
+				ArrayHelper::toInteger($array['categories']);
 				$this->categories = implode("|",$array['categories']);
 			}
 		}
 		if (key_exists('calendars', $array)) {
 			if($array['calendars']=='all' || $array['calendars']=='none') $this->calendars = $array['calendars'];
 			else if (is_array($array['calendars'])){
-				JArrayHelper::toInteger($array['calendars']);
+				ArrayHelper::toInteger($array['calendars']);
 				$this->calendars = implode("|",$array['calendars']);
 			}
 		}